src/requirements.txt
aniso8601==8.1.1
asgiref==3.3.4
boto3==1.17.82
botocore==1.20.82
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
cryptography==3.4.7
Django==3.2.3
django-appconf==1.0.4
django-autocomplete-light==3.8.2
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-hosts==4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-nested-admin==3.3.3
django-registration==3.1.2
django-split-settings==1.0.1
django-storages==1.11.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.4
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
greenlet==1.1.0
idna==2.10
jmespath==0.10.0
msgpack==1.0.2
mypy==0.812
mypy-extensions==0.4.3
pilkit==2.0
Pillow==8.2.0
promise==2.3
psycopg2==2.8.6
pycparser==2.20
pynvim==0.4.3
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytz==2021.1
requests==2.25.1
Rx==1.6.1
s3transfer==0.4.2
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
